Update README.md
This project is a soul message from Nickson—a wake-up call to stop faking and start living. It’s more than a webpage; it’s a reminder that life is beautiful, short, and real. Inspired by experiences beyond the surface, this project aims to awaken authenticity and purpose in others through raw truth and self-expression.